+++ masterpiece-web/src/app/app.module.ts
@@ -1,12 +1,12 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
-import { SharedModule } from './shared/shared.module'
+import { SharedModule } from './shared/shared.module';
+import { CoreModule } from './core/core.module';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FlexLayoutModule } from '@angular/flex-layout';
-import { MaterialModule } from './material/material.module';
 import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
 import { TranslateHttpLoader } from '@ngx-translate/http-loader';
 import { HttpClientModule, HttpClient, HTTP_INTERCEPTORS } from '@angular/common/http';
@@ -26,9 +26,9 @@ import { DatePipe } from '@angular/common';
     BrowserModule,
     AppRoutingModule,
     SharedModule,
+    CoreModule,
     BrowserAnimationsModule,
     FlexLayoutModule,
-    MaterialModule,
     MatDialogModule,
     HttpClientModule,
     TranslateModule.forRoot({
