+++ masterpiece-web/src/app/shared/layout/layout.component.spec.ts
