+++ masterpiece-web/src/app/shared/shared.module.ts
@@ -2,21 +2,38 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NotFoundComponent } from './page-not-found/not-found.component';
 import { BtnCellRenderer } from './btn-cell-renderer.component';
-import { MatDialogModule, MatInputModule ,MatFormFieldModule } from '@angular/material';
-import { DateTimeDialogComponentComponent } from './modals/date-time-dialog-component/date-time-dialog-component.component';
+import { MatToolbarModule, MatIconModule, MatSidenavModule, MatListModule, MatButtonModule, MatInputModule, MatFormFieldModule, MatDialogModule, MatMenuModule } from '@angular/material';import { DateTimeDialogComponentComponent } from './modals/date-time-dialog-component/date-time-dialog-component.component';
 import { ReactiveFormsModule, FormsModule } from '@angular/forms';
 import { MatDatepickerModule } from '@angular/material/datepicker';
 import { NgxMatDatetimePickerModule, NgxMatTimepickerModule, NgxMatNativeDateModule } from '@angular-material-components/datetime-picker';
-import { TranslateModule } from '@ngx-translate/core';
+import { TranslateLoader, TranslateModule } from '@ngx-translate/core';
+import { NavbarComponent } from './navbar/navbar.component';
+import { LayoutComponent } from './layout/layout.component';
+import { FlexLayoutModule } from '@angular/flex-layout';
+import { NavbarRoutingModule } from './navbar-routing.module';
+import { HttpClient } from '@angular/common/http';
+import { TranslateHttpLoader } from '@ngx-translate/http-loader';
+import { RouterModule } from '@angular/router'
+
 @NgModule({
   declarations: [
     NotFoundComponent,
     BtnCellRenderer,
-    DateTimeDialogComponentComponent
+    DateTimeDialogComponentComponent,
+    LayoutComponent,
+    NavbarComponent
   ],
   imports: [
+    RouterModule,
+    FlexLayoutModule,
     CommonModule,
     MatDialogModule,
+    MatToolbarModule,
+    MatSidenavModule,
+    MatListModule,
+    MatButtonModule,
+    MatIconModule,
+    MatMenuModule,
     ReactiveFormsModule,
     FormsModule,
     MatInputModule,
@@ -25,13 +42,28 @@ import { TranslateModule } from '@ngx-translate/core';
     NgxMatDatetimePickerModule, 
     NgxMatTimepickerModule, 
     NgxMatNativeDateModule,
-    TranslateModule
+    TranslateModule,
+    MatIconModule,
+    NavbarRoutingModule,
+    TranslateModule.forChild({
+      loader: {
+        provide: TranslateLoader,
+        useFactory: HttpLoaderFactory,
+        deps: [HttpClient]
+      }
+    })
   ],
   exports: [
     BtnCellRenderer,
+    LayoutComponent,
+    NavbarComponent
   ],
   entryComponents: [
     DateTimeDialogComponentComponent
   ]
 })
 export class SharedModule { }
+
+export function HttpLoaderFactory(http: HttpClient) {
+  return new TranslateHttpLoader(http);
+}
